requirements/py2.txt
# Import Base Requirements
-r _base.txt

ldap==1.0.2